requirements.txt
earthengine-api>=0.1.350
geopandas>=0.12.0
pandas>=1.5.0
numpy>=1.21.0
rasterio>=1.3.0
shapely>=1.8.0
scikit-learn>=1.1.0
matplotlib>=3.5.0
seaborn>=0.11.0
folium>=0.14.0
plotly>=5.0.0
tqdm>=4.64.0
pyyaml>=6.0
requests>=2.28.0
python-dateutil>=2.8.0